LayoutTests/editing/inserting/insert-paragraph-inside-form.html

<!DOCTYPE html>
<html>
<body>
<div id="test" contenteditable><form>hello<br>world</form></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests inserting a paragraph inside a form element. We should do clone the form element.')

var form = document.querySelector('form');
getSelection().collapse(form, 1);
document.execCommand('InsertParagraph', false, null);

Markup.dump('test');

</script>
</body>
</html>
