third_party/blink/web_tests/platform/generic/editing/inserting/insert-paragraph-inside-form-expected.txt

This tests inserting a paragraph inside a form element. We should do clone the form element.
| <form>
|   "hello"
|   <br>
|   <#selection-caret>
|   <br>
|   "world"
